MNT: move logic to different module
